make EIP2612 Check Better, move around types a bit
