+++ background/types.ts
@@ -67,19 +67,4 @@ export type EIP712TypedData<T = Record<string, unknown>> = {
   primaryType: string
 }
 
-type EIP2612Message = {
-  owner: HexString
-  spender: HexString
-  value: number
-  nonce: number
-  deadline: number
-}
-
-export type EIP2612TypedData = {
-  domain: EIP712DomainType
-  types: Record<string, TypedDataField[]>
-  message: EIP2612Message
-  primaryType: "Permit"
-}
-
 export type EIP191Data = string
